+++ nextstep-mvc/src/main/java/nextstep/mvc/tobe/core/ComponentScanner.java
@@ -1,4 +1,4 @@
-package nextstep.mvc.tobe;
+package nextstep.mvc.tobe.core;
 
 public interface ComponentScanner {
     Object scan();
