+++ nextstep-mvc/src/main/java/nextstep/mvc/tobe/core/HandlerKey.java
@@ -1,4 +1,4 @@
-package nextstep.mvc.tobe;
+package nextstep.mvc.tobe.core;
 
 import nextstep.web.annotation.RequestMethod;
