+++ slipp/src/test/java/slipp/controller/CreateUserControllerTest.java
@@ -1,6 +1,6 @@
 package slipp.controller;
 
-import nextstep.mvc.tobe.RequestMappingHandlerMapping;
+import nextstep.mvc.tobe.core.RequestMappingHandlerMapping;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
